# contract/openapi.yaml
...
components:
  schemas:
    AccessGranteeType:
      enum:
      - "USER"
      - "WORKSPACE"
      type: "string"
...
    AddResourceAccessGrantRequest:
      properties:
        granteeType:
          $ref: "#/components/schemas/AccessGranteeType"
          description: "USER는 지정된 사용자 한 명, WORKSPACE은 소유 워크스페이스 전체"
        role:
          $ref: "#/components/schemas/ResourceRole"
          description: "부여할 등급. 워크스페이스 전체 항목에는 MEMBER 또는 VIEWER만 지정할 수 있습니다."
        userId:
          description: "대상 사용자 id. granteeType이 USER일 때만 보내며, 소유 워크스페이스의 구성원이어야 합니\
            다."
          format: "uuid"
          type:
          - "string"
          - "null"
      required:
      - "granteeType"
      - "role"
      type: "object"
...
    CreateOrgRequest:
      properties:
        description:
          maxLength: 500
          minLength: 0
          type:
          - "string"
          - "null"
        name:
          maxLength: 100
          minLength: 0
          type: "string"
      required:
      - "name"
      type: "object"
...
    DriftFindingKind:
      enum:
      - "MISSING_IN_PROXMOX"
      - "UNMANAGED_GUEST"
      - "SPEC_MISMATCH"
      type: "string"
    DriftFindingResponse:
      properties:
        detail: {}
        firstSeenAt:
          format: "date-time"
          type: "string"
        id:
          format: "uuid"
          type: "string"
        kind:
          $ref: "#/components/schemas/DriftFindingKind"
        lastSeenAt:
          format: "date-time"
          type: "string"
        nodeName:
          type:
          - "string"
          - "null"
        proxmoxVmid:
          format: "int32"
          type:
          - "integer"
          - "null"
        resolutionNote:
          type:
          - "string"
          - "null"
        resolvedAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        resolvedByEmail:
          type:
          - "string"
          - "null"
        resolvedById:
          format: "uuid"
          type:
          - "string"
          - "null"
        status:
          $ref: "#/components/schemas/DriftFindingStatus"
        summary:
          type: "string"
        vmId:
          format: "uuid"
          type:
          - "string"
          - "null"
        vmName:
          type:
          - "string"
          - "null"
      required:
      - "firstSeenAt"
      - "id"
      - "kind"
      - "lastSeenAt"
      - "status"
      - "summary"
      type: "object"
    DriftFindingStatus:
      enum:
      - "OPEN"
      - "RESOLVED"
      type: "string"
    FieldValidationError:
      description: "검증 실패 필드 한 건 (422 응답의 errors[] 원소)"
      properties:
        field:
          description: "실패한 요청 필드 이름"
          type: "string"
        message:
          description: "한국어 오류 메시지"
          type: "string"
      required:
      - "field"
      - "message"
      type: "object"
...
    Grantee:
      description: "접근 권한을 가진 사용자"
      properties:
        email:
          type: "string"
        name:
          type: "string"
        userId:
          format: "uuid"
          type: "string"
      required:
      - "email"
      - "name"
      - "userId"
      type: "object"
...
    Membership:
      properties:
        role:
          $ref: "#/components/schemas/WorkspaceMemberRole"
        workspaceId:
          format: "uuid"
          type: "string"
        workspaceKind:
          $ref: "#/components/schemas/WorkspaceKind"
        workspaceName:
          type: "string"
      required:
      - "role"
      - "workspaceId"
      - "workspaceKind"
      - "workspaceName"
      type: "object"
...
    OrgDetailResponse:
      properties:
        createdAt:
          format: "date-time"
          type: "string"
        description:
          type:
          - "string"
          - "null"
        hidden:
          type: "boolean"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
        status:
          $ref: "#/components/schemas/OrgStatus"
      required:
      - "createdAt"
      - "hidden"
      - "id"
      - "name"
      - "status"
      type: "object"
...
    OrgStatus:
      enum:
      - "ACTIVE"
      - "DISABLED"
      type: "string"
    OrgSummaryResponse:
      properties:
        description:
          type:
          - "string"
          - "null"
        hidden:
          type: "boolean"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
        status:
          $ref: "#/components/schemas/OrgStatus"
      required:
      - "hidden"
      - "id"
      - "name"
      - "status"
      type: "object"
...
    PageResponseDriftFindingResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/DriftFindingResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      required:
      - "content"
      - "page"
      - "size"
      - "totalElements"
      - "totalPages"
      type: "object"
...
    Problem:
      description: "RFC 9457 오류 응답 + Pickle 확장(code, errors). 모든 비정상 응답은 이 형태로 반환됩\
        니다."
      properties:
        code:
          description: "기계 판독용 안정 오류 코드 (클라이언트 분기 기준)"
          type: "string"
        detail:
          description: "이 발생 건에 대한 설명"
          type: "string"
        errors:
          description: "검증 실패 상세 (422에서만 존재)"
          items:
            $ref: "#/components/schemas/FieldValidationError"
          type: "array"
        instance:
          description: "발생 경로"
          type: "string"
        status:
          description: "HTTP 상태 코드"
          format: "int32"
          type: "integer"
        title:
          description: "사람이 읽는 짧은 제목"
          type: "string"
        type:
          description: "문제 유형 URI (기본 about:blank)"
          type: "string"
      required:
      - "title"
      - "status"
      - "code"
      type: "object"
...
    ResolveDriftFindingRequest:
      properties:
        note:
          maxLength: 2000
          minLength: 0
          type: "string"
      type: "object"
...
    ResourceAccessGrantView:
      description: "VM 접근 권한 한 건. 대상은 지정된 사용자 한 명이거나 소유 워크스페이스 전체입니다."
      properties:
        createdAt:
          format: "date-time"
          type: "string"
        granteeType:
          $ref: "#/components/schemas/AccessGranteeType"
          description: "대상 종류 — USER는 지정된 사용자, WORKSPACE은 소유 워크스페이스 전체"
        id:
          format: "uuid"
          type: "string"
        role:
          $ref: "#/components/schemas/ResourceRole"
          description: "이 대상이 이 VM에서 갖는 등급"
        user:
          anyOf:
          - $ref: "#/components/schemas/Grantee"
          - type: "null"
          description: "대상 사용자. 워크스페이스 전체 항목이면 null입니다."
      required:
      - "createdAt"
      - "granteeType"
      - "id"
      - "role"
      type: "object"
    ResourceAccessListResponse:
      description: "리소스 접근 권한 목록과, 그 목록이 어느 리소스의 것인지 알려 주는 최소 정보"
      properties:
        grants:
          description: "접근 권한 항목"
          items:
            $ref: "#/components/schemas/ResourceAccessGrantView"
          type: "array"
        resource:
          $ref: "#/components/schemas/ResourceBrief"
          description: "이 목록이 속한 리소스"
      required:
      - "grants"
      - "resource"
      type: "object"
    ResourceBrief:
      description: "접근 권한이 없는 사람에게도 보이는 범위의 리소스 정보"
      properties:
        displayName:
          description: "표시명. 지정되지 않았으면 null입니다."
          type:
          - "string"
          - "null"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
        status:
          description: "리소스 종류의 상태값. 종류마다 어휘가 다릅니다."
          type: "string"
        type:
          $ref: "#/components/schemas/ResourceType"
        workspaceId:
          format: "uuid"
          type: "string"
        workspaceName:
          type: "string"
      required:
      - "id"
      - "name"
      - "status"
      - "type"
      - "workspaceId"
      - "workspaceName"
      type: "object"
    ResourceRole:
      enum:
      - "OWNER"
      - "EDITOR"
      - "MEMBER"
      - "VIEWER"
      type: "string"
...
    ResourceType:
      enum:
      - "VM"
      type: "string"
...
    UpdateOrgRequest:
      properties:
        description:
          maxLength: 500
          minLength: 0
          type:
          - "string"
          - "null"
        hidden:
          type: "boolean"
        name:
          maxLength: 100
          minLength: 0
          type: "string"
        status:
          $ref: "#/components/schemas/OrgStatus"
      type: "object"
...
    UpdateResourceAccessGrantRequest:
      properties:
        role:
          $ref: "#/components/schemas/ResourceRole"
          description: "새 등급. 워크스페이스 전체 항목에는 MEMBER 또는 VIEWER만 지정할 수 있습니다."
      required:
      - "role"
      type: "object"
    UpdateUserAdminRequest:
      properties:
        orgId:
          format: "uuid"
          type:
          - "string"
          - "null"
        role:
          $ref: "#/components/schemas/UserRole"
      type: "object"
...
    UserAdminDetailResponse:
      properties:
        activeVmCount:
          format: "int32"
          type: "integer"
        createdAt:
          format: "date-time"
          type: "string"
        disabledAt:
          format: "date-time"
          type:
          - "string"
          - "null"
        disabledReason:
          type:
          - "string"
          - "null"
        email:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        memberships:
          items:
            $ref: "#/components/schemas/Membership"
          type: "array"
        mfaEnabled:
          type: "boolean"
        name:
          type: "string"
        orgId:
          format: "uuid"
          type:
          - "string"
          - "null"
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        statusChanges:
          items:
            $ref: "#/components/schemas/UserStatusChangeResponse"
          type: "array"
        withdrawnAt:
          format: "date-time"
          type:
          - "string"
          - "null"
      required:
      - "activeVmCount"
      - "createdAt"
      - "email"
      - "id"
      - "memberships"
      - "mfaEnabled"
      - "name"
      - "role"
      - "status"
      - "statusChanges"
      type: "object"
...
    UserRole:
      enum:
      - "USER"
      - "ORG_MANAGER"
      - "ORG_ADMIN"
      - "SYS_MANAGER"
      - "SYS_ADMIN"
      type: "string"
    UserStatus:
      enum:
      - "PENDING_VERIFICATION"
      - "ACTIVE"
      - "DISABLED"
      - "WITHDRAWN"
      type: "string"
    UserStatusChangeResponse:
      properties:
        actorEmail:
          type:
          - "string"
          - "null"
        actorId:
          format: "uuid"
          type:
          - "string"
          - "null"
        changedAt:
          format: "date-time"
          type: "string"
        fromStatus:
          $ref: "#/components/schemas/UserStatus"
        reason:
          type:
          - "string"
          - "null"
        toStatus:
          $ref: "#/components/schemas/UserStatus"
      required:
      - "changedAt"
      - "fromStatus"
      - "toStatus"
      type: "object"
    UserSummaryResponse:
      properties:
        email:
          type: "string"
        id:
          format: "uuid"
          type: "string"
        name:
          type: "string"
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - "email"
      - "id"
      - "name"
      - "role"
      type: "object"
...
    WorkspaceKind:
      enum:
      - "PERSONAL"
      - "TEAM"
      - "PROJECT"
      type: "string"
...
    WorkspaceMemberRole:
      enum:
      - "OWNER"
      - "MEMBER"
      type: "string"
...
  securitySchemes:
    bearerAuth:
      bearerFormat: "JWT"
      description: "액세스 토큰 (JWT HS256, 15분 만료). 클레임: sub(계정 공개 식별자 UUID), role, token_version.\
        \ 비밀번호 변경·계정 비활성화 시 token_version이 올라가 기존 토큰이 즉시 무효화됩니다. 리프레시 토큰은 보안 스킴이 아\
        니라 __Host-pickle_refresh httpOnly 쿠키로만 오갑니다."
      scheme: "bearer"
      type: "http"
info:
  description: "부산대학교 클라우드 플랫폼 Pickle의 REST API. 인증은 JWT Bearer, 오류 응답은 RFC 9457 problem+json(Problem\
    \ 스키마)을 따릅니다."
  title: "Pickle API"
  version: "0.38.0"
openapi: "3.1.0"
paths:
...
  /admin/drift-findings:
    get:
      operationId: "listDriftFindings"
      parameters:
      - in: "query"
        name: "status"
        required: false
        schema:
          $ref: "#/components/schemas/DriftFindingStatus"
      - in: "query"
        name: "kind"
        required: false
        schema:
          $ref: "#/components/schemas/DriftFindingKind"
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          minimum: 0
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          maximum: 100
          minimum: 1
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDriftFindingResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-drift-controller"
  /admin/drift-findings/{findingId}/resolve:
    post:
      operationId: "resolveDriftFinding"
      parameters:
      - in: "path"
        name: "findingId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDriftFindingRequest"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DriftFindingResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-drift-controller"
...
  /admin/orgs:
    get:
      operationId: "listAdminOrgs"
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: "#/components/schemas/OrgDetailResponse"
                type: "array"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-controller"
    post:
      operationId: "createOrg"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrgDetailResponse"
          description: "Created"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-controller"
  /admin/orgs/{orgId}:
    patch:
      operationId: "updateOrg"
      parameters:
      - in: "path"
        name: "orgId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrgDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-controller"
...
  /admin/users/{userId}:
    get:
      operationId: "getUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserAdminDetailResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-user-controller"
    patch:
      operationId: "updateUser"
      parameters:
      - in: "path"
        name: "userId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserAdminRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSummaryResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "admin-controller"
...
  /orgs:
    get:
      operationId: "listOrgs"
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: "#/components/schemas/OrgSummaryResponse"
                type: "array"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      tags:
      - "org-controller"
...
  /vms/{vmId}/access:
    get:
      description: "이 VM의 접근 권한 전체와, 그 목록이 어느 VM의 것인지 알려 주는 최소 정보입니다. VM 소유자와 워크스페\
        이스 소유자만 볼 수 있습니다."
      operationId: "listVmAccessGrants"
      parameters:
      - in: "path"
        name: "vmId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceAccessListResponse"
          description: "OK"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      summary: "접근 권한 목록"
      tags:
      - "vm-access"
    post:
      description: "지정한 사용자 또는 소유 워크스페이스 전체에 이 VM의 접근 권한을 부여합니다. 사용자는 이 VM을 소유한 워크\
        스페이스의 구성원이어야 하고, 워크스페이스 전체에는 참여자·열람자까지만 부여할 수 있습니다."
      operationId: "addVmAccessGrant"
      parameters:
      - in: "path"
        name: "vmId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddResourceAccessGrantRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceAccessGrantView"
          description: "Created"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      summary: "접근 권한 부여"
      tags:
      - "vm-access"
  /vms/{vmId}/access/{grantId}:
    delete:
      description: "회수해도 이미 열람한 초기 비밀번호와 이미 열려 있는 SSH 세션은 회수되지 않습니다. 필요하면 비밀번호를 재생\
        성해 주세요."
      operationId: "removeVmAccessGrant"
      parameters:
      - in: "path"
        name: "vmId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - in: "path"
        name: "grantId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      responses:
        "204":
          description: "No Content"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      summary: "접근 권한 회수"
      tags:
      - "vm-access"
    patch:
      operationId: "updateVmAccessGrant"
      parameters:
      - in: "path"
        name: "vmId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - in: "path"
        name: "grantId"
        required: true
        schema:
          format: "uuid"
          type: "string"
      - description: "재인증(sudo-mode) 토큰 — POST /auth/reverify가 발급 (10분 유효, 다회용). 없\
          거나 만료·무효면 403 REAUTH_REQUIRED."
        in: "header"
        name: "X-Reauth-Token"
        required: false
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceAccessGrantRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceAccessGrantView"
          description: "OK"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "재인증 필요 — 유효한 X-Reauth-Token 없음 (`REAUTH_REQUIRED`)"
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "오류 — 상태 코드와 무관하게 Problem 형태"
      summary: "접근 권한 등급 변경"
      tags:
      - "vm-access"
...
security:
- bearerAuth: []
servers:
- url: "/api/v1"
...